+++ src/index.ts
@@ -5,8 +5,6 @@ import CommandBuild from './commands/build'
 import CommandServe from './commands/serve'
 import CommandLint from './commands/lint'
 import CommandFormat from './commands/format'
-import CommandTypesCreate from './commands/types-create'
-import CommandTypesSync from './commands/types-sync'
 
 import TemplateBase from './templates/base'
 import TemplateVue from './templates/vue'
@@ -18,8 +16,6 @@ export default class FactoryWeb extends Factory {
   commands = [
     new CommandBuild(this),
     new CommandServe(this),
-    new CommandTypesCreate(this),
-    new CommandTypesSync(this),
     new CommandLint(this),
     new CommandFormat(this)
   ]
@@ -33,12 +29,4 @@ export default class FactoryWeb extends Factory {
   }
 }
 
-export {
-  CommandBuild,
-  CommandServe,
-  CommandTypesCreate,
-  CommandTypesSync,
-  TemplateBase,
-  TemplateVue,
-  TemplateReact
-}
+export { CommandBuild, CommandServe, TemplateBase, TemplateVue, TemplateReact }